Added a timeout parameter to prog2file() which defaults to 30 seconds. This uses alarm().  Also fixed some bugs in freshenup.pl and daily_archive.pl.

+++ Slash/Utility/System/System.pm
@@ -1,7 +1,7 @@
 # This code is a part of Slash, and is released under the GPL.
 # Copyright 1997-2002 by Open Source Development Network. See README
 # and COPYING for more information, or see http://slashcode.com/.
-# $Id: System.pm,v 1.12 2002/06/14 00:18:39 jamie Exp $
+# $Id: System.pm,v 1.13 2002/06/19 16:57:42 jamie Exp $
 
 package Slash::Utility::System;
 
@@ -39,7 +39,7 @@ use Symbol 'gensym';
 use base 'Exporter';
 use vars qw($VERSION @EXPORT @EXPORT_OK);
 
-($VERSION) = ' $Revision: 1.12 $ ' =~ /\$Revision:\s+([^\s]+)/;
+($VERSION) = ' $Revision: 1.13 $ ' =~ /\$Revision:\s+([^\s]+)/;
 @EXPORT	   = qw(
 	bulkEmail
 	doEmail
@@ -321,15 +321,31 @@ sub prog2file {
 
 	# Two ways of handling data from child programs yet we maintain
 	# backwards compatibility.
-	# Need an alarm() around these.
-	if (! $handle_err) {
-		$data = `$exec`;
-	} else {
-		my($errfh, $errfile) = tempfile();
-		$data = `$exec 2>$errfile`;
-		$stderr_text = join '', <$errfh>;
-		close $errfh;
-		unlink $errfile;
+	# Passing "timeout" as a field to $options does what you'd think.
+	# A timeout of 0 means "never time out".  30 seconds is default.
+	my $timeout = 30;
+	$timeout = $options->{timeout} if defined($options->{timeout});
+	my($errfh, $errfile) = (undef, undef);
+	eval {
+		local $SIG{ALRM} = sub { die "timeout" };
+		alarm $timeout if $timeout;
+		if (!$handle_err) {
+			$data = `$exec`;
+			alarm 0 if $timeout;
+		} else {
+			($errfh, $errfile) = tempfile();
+			$data = `$exec 2>$errfile`;
+			alarm 0 if $timeout;
+			$stderr_text = join '', <$errfh>;
+			close $errfh; $errfh = undef;
+			unlink $errfile; $errfile = undef;
+		}
+	};
+	my $success_str = "";
+	if ($timeout && $@ && $@ =~ /timeout/) {
+		$success_str = " TIMEOUT_HIT";
+		close $errfh if $errfh;
+		unlink $errfile if $errfile;
 	}
 	my $bytes = length $data;
 
@@ -354,7 +370,7 @@ sub prog2file {
 
 	my($command_base) = $command =~ m{([^/]+)$};
 	$command_base ||= $command;
-	my $success_str = $success ? "" : " FAILED:$err_str";
+	$success_str .= $success ? "" : " FAILED:$err_str";
 	$success_str =~ s/\s+/ /g; chomp $success_str;
 
 	if ($verbosity >= 2) {
@@ -399,4 +415,4 @@ Slash(3), Slash::Utility(3).
 
 =head1 VERSION
 
-$Id: System.pm,v 1.12 2002/06/14 00:18:39 jamie Exp $
+$Id: System.pm,v 1.13 2002/06/19 16:57:42 jamie Exp $
